upgraded sub-task 2 in task 2757, reformatted code
