+++ tes-adapter/src/main/java/com/epam/pipeline/adaptor/applications/TesAdaptorsApplication.java
@@ -1,0 +1,12 @@
+package com.epam.pipeline.adaptor.applications;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TesAdaptorsApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(TesAdaptorsApplication.class, args);
+    }
+}
