+++ tes-adapter/src/main/resources/templates/OK.mustache
@@ -1,0 +1,7 @@
+<html>
+<body>
+<div>
+    hello
+</div>
+</body>
+</html>
